Add a cli parser
